Rename processor/test to processortest
