+++ sdk/metric/controller/pull/pull_test.go
@@ -28,7 +28,7 @@ import (
 	export "go.opentelemetry.io/otel/sdk/export/metric"
 	"go.opentelemetry.io/otel/sdk/metric/controller/pull"
 	controllerTest "go.opentelemetry.io/otel/sdk/metric/controller/test"
-	"go.opentelemetry.io/otel/sdk/metric/processor/test"
+	"go.opentelemetry.io/otel/sdk/metric/processor/processortest"
 	selector "go.opentelemetry.io/otel/sdk/metric/selector/simple"
 )
 
@@ -46,7 +46,7 @@ func TestPullNoCache(t *testing.T) {
 	counter.Add(ctx, 10, kv.String("A", "B"))
 
 	require.NoError(t, puller.Collect(ctx))
-	records := test.NewOutput(label.DefaultEncoder())
+	records := processortest.NewOutput(label.DefaultEncoder())
 	require.NoError(t, puller.ForEach(export.CumulativeExporter, records.AddRecord))
 
 	require.EqualValues(t, map[string]float64{
@@ -56,7 +56,7 @@ func TestPullNoCache(t *testing.T) {
 	counter.Add(ctx, 10, kv.String("A", "B"))
 
 	require.NoError(t, puller.Collect(ctx))
-	records = test.NewOutput(label.DefaultEncoder())
+	records = processortest.NewOutput(label.DefaultEncoder())
 	require.NoError(t, puller.ForEach(export.CumulativeExporter, records.AddRecord))
 
 	require.EqualValues(t, map[string]float64{
@@ -80,7 +80,7 @@ func TestPullWithCache(t *testing.T) {
 	counter.Add(ctx, 10, kv.String("A", "B"))
 
 	require.NoError(t, puller.Collect(ctx))
-	records := test.NewOutput(label.DefaultEncoder())
+	records := processortest.NewOutput(label.DefaultEncoder())
 	require.NoError(t, puller.ForEach(export.CumulativeExporter, records.AddRecord))
 
 	require.EqualValues(t, map[string]float64{
@@ -91,7 +91,7 @@ func TestPullWithCache(t *testing.T) {
 
 	// Cached value!
 	require.NoError(t, puller.Collect(ctx))
-	records = test.NewOutput(label.DefaultEncoder())
+	records = processortest.NewOutput(label.DefaultEncoder())
 	require.NoError(t, puller.ForEach(export.CumulativeExporter, records.AddRecord))
 
 	require.EqualValues(t, map[string]float64{
@@ -103,7 +103,7 @@ func TestPullWithCache(t *testing.T) {
 
 	// Re-computed value!
 	require.NoError(t, puller.Collect(ctx))
-	records = test.NewOutput(label.DefaultEncoder())
+	records = processortest.NewOutput(label.DefaultEncoder())
 	require.NoError(t, puller.ForEach(export.CumulativeExporter, records.AddRecord))
 
 	require.EqualValues(t, map[string]float64{
